modification to account ui

+++ app/src/main/java/com/example/market/ui/search/SearchFragment.kt
@@ -1,5 +1,7 @@
 package com.example.market.ui.search
 
+import android.content.Intent
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -7,17 +9,21 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.SearchView
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import com.example.market.R
+import com.example.market.auth.AuthActivity
 import com.example.market.data.pojo.Product
 import com.example.market.databinding.FragmentSearchBinding
+import com.example.market.utils.Constants
 import kotlin.math.abs
 import com.example.market.utils.NetworkResult
 import com.google.android.material.slider.Slider
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 import kotlin.math.roundToInt
 
 @AndroidEntryPoint
@@ -27,6 +33,9 @@ class SearchFragment : Fragment() {
     private val binding get() = _binding!!
 
     private val viewModel: SearchViewModel by viewModels()
+
+    @Inject
+    lateinit var sharedPreferences: SharedPreferences
 
     private val searchAdapter =
         SearchAdapter(object : SearchAdapter.ProductClickListener {
@@ -39,12 +48,16 @@ class SearchFragment : Fragment() {
             }
 
             override fun onFavouriteClicked(product: Product) {
-                if (product.isFavourite) {
-                    viewModel.deleteFavourite(product)
+                if (sharedPreferences.getBoolean(Constants.IS_Logged, false)) {
+                    if (product.isFavourite) {
+                        viewModel.deleteFavourite(product)
+                    } else {
+                        viewModel.addFavourite(product)
+                    }
+                    product.isFavourite = !product.isFavourite
                 } else {
-                    viewModel.addFavourite(product)
+                    showAlertDialog()
                 }
-                product.isFavourite = !product.isFavourite
             }
         })
 
@@ -153,6 +166,26 @@ class SearchFragment : Fragment() {
             layoutManager =
                 GridLayoutManager(requireContext(), 2, GridLayoutManager.VERTICAL, false)
         }
+    }
+
+    private fun showAlertDialog() {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Error")
+        builder.setMessage("You Must login first.")
+        builder.setIcon(android.R.drawable.ic_dialog_alert)
+        builder.setPositiveButton(resources.getString(R.string.OK)) { _, _ ->
+            val i = Intent(requireActivity(), AuthActivity::class.java)
+            i.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
+            startActivity(i)
+            activity?.finish()
+        }
+        builder.setNegativeButton(resources.getString(R.string.cancel)) { dialogInterface, _ ->
+            dialogInterface.dismiss()
+        }
+
+        val alertDialog: AlertDialog = builder.create()
+        alertDialog.setCancelable(false)
+        alertDialog.show()
     }
 
     override fun onDestroyView() {
